+++ src/twavetents.hpp
@@ -105,7 +105,7 @@ namespace ngcomp
     int fosystem = 0;
     double timeshift = 0;
     int nbasis;
-    const int nsimd = SIMD<double>::Size ();
+    const size_t nsimd = SIMD<double>::Size ();
     static constexpr ELEMENT_TYPE eltyp
         = (D == 3) ? ET_TET : ((D == 2) ? ET_TRIG : ET_SEGM);
